Avoid most joins with `solutions`
They used huge amounts of memory. The current approach, where we keep
separate data.tables for steady states etc, leads to messier code, but
it is likely to use less memory than adding a number of columns to
solutions that are const for each rep. That said, I've not tested any of
the above claims.